Split OpenAPI config from apiconfig
